+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/AbstractTigerRouteCallback.java
@@ -7,17 +7,18 @@ package de.gematik.test.tiger.proxy;
 import static org.mockserver.model.Header.header;
 import static org.mockserver.model.HttpOverrideForwardedRequest.forwardOverriddenRequest;
 import de.gematik.rbellogger.data.RbelElement;
-import de.gematik.rbellogger.data.facet.RbelHttpResponseFacet;
-import de.gematik.rbellogger.data.facet.RbelMessageTimingFacet;
-import de.gematik.rbellogger.data.facet.RbelUriFacet;
-import de.gematik.rbellogger.data.facet.RbelUriParameterFacet;
+import de.gematik.rbellogger.data.facet.*;
 import de.gematik.test.tiger.common.data.config.tigerProxy.TigerRoute;
+import de.gematik.test.tiger.proxy.certificate.TlsFacet;
 import de.gematik.test.tiger.proxy.data.TracingMessagePairFacet;
 import de.gematik.test.tiger.proxy.exceptions.TigerProxyModificationException;
+import java.security.cert.Certificate;
+import java.security.cert.CertificateEncodingException;
 import java.time.ZonedDateTime;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -174,6 +175,9 @@ public abstract class AbstractTigerRouteCallback implements ExpectationForwardAn
                         .request(request)
                         .build());
 
+                parseCertificateChainIfPresent(req, request)
+                    .ifPresent(request::addFacet);
+
                 getTigerProxy().triggerListener(request);
                 getTigerProxy().triggerListener(response);
             } catch (RuntimeException e) {
@@ -182,6 +186,36 @@ public abstract class AbstractTigerRouteCallback implements ExpectationForwardAn
             }
         }
         return resp.withBody(resp.getBodyAsRawBytes());
+    }
+
+    private Optional<RbelFacet> parseCertificateChainIfPresent(HttpRequest httpRequest, RbelElement message) {
+        if (httpRequest.getClientCertificateChain() == null ||
+            httpRequest.getClientCertificateChain().isEmpty()) {
+            return Optional.empty();
+        }
+        RbelElement certificateChainElement = new RbelElement(null, message);
+        certificateChainElement.addFacet(RbelListFacet.builder()
+            .childNodes(httpRequest.getClientCertificateChain().stream()
+                .map(X509Certificate::getCertificate)
+                .map(cert -> mapToRbelElement(cert, message))
+                .collect(Collectors.toList()))
+            .build());
+
+        return Optional.of(
+            TlsFacet.builder()
+                .clientCertificateChain(certificateChainElement)
+                .build()
+        );
+    }
+
+    private RbelElement mapToRbelElement(Certificate certificate, RbelElement parentNode) {
+        try {
+            final RbelElement certificateNode = new RbelElement(certificate.getEncoded(), parentNode);
+            getTigerProxy().getRbelLogger().getRbelConverter().convertElement(certificateNode);
+            return certificateNode;
+        } catch (CertificateEncodingException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     boolean shouldLogTraffic() {